PIM-7660: put label logic into the query instead of the read model
